refactor: information 컴포넌트 top, bottom 컴포넌트로 분리
